pipeline/requirements.txt
contourpy==1.0.7
cycler==0.11.0
fonttools==4.39.4
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.24.3
packaging==23.1
pandas==1.5.3
Pillow==9.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
threadpoolctl==3.1.0
tzdata==2023.3
xgboost==1.7.5
streamlit==1.22.0
boto3==1.26.142
typer==0.9.0
datasets==2.12.0